重构组件结构
