+++ .gitignore
@@ -1,11 +1,13 @@
-*.iml
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
 .DS_Store
+.idea/
+*.iml
+*.apk
+
+/local.properties
+
 /build
 /captures
 .externalNativeBuild
-/app/
-app/libs/
+
+.gradle/
+gradle/
