+++ app/src/main/java/com/recyclerview/loadmore/sample/fragment/FragmentThree.java
@@ -1,0 +1,22 @@
+package com.recyclerview.loadmore.sample.fragment;
+
+import com.recyclerview.loadmore.sample.R;
+import com.recyclerview.loadmore.sample.base.BaseFrgment;
+
+/**
+ * zhangyao
+ * 16/9/24
+ * zhangyao@jiandanxinli.com
+ */
+
+public class FragmentThree extends BaseFrgment {
+    @Override
+    protected int getLayoutId() {
+        return R.layout.fragment_three;
+    }
+
+    @Override
+    public void init() {
+
+    }
+}
