+++ app/src/main/java/com/recyclerview/loadmore/sample/fragment/FragmentTwo.java
@@ -1,0 +1,22 @@
+package com.recyclerview.loadmore.sample.fragment;
+
+import com.recyclerview.loadmore.sample.R;
+import com.recyclerview.loadmore.sample.base.BaseFrgment;
+
+/**
+ * zhangyao
+ * 16/9/24
+ * zhangyao@jiandanxinli.com
+ */
+
+public class FragmentTwo extends BaseFrgment {
+    @Override
+    protected int getLayoutId() {
+        return R.layout.fragment_two;
+    }
+
+    @Override
+    public void init() {
+
+    }
+}
