--- app/src/main/java/com/recyclerviewmore/fragment/FragmentTwo.java
@@ -1,22 +1,0 @@
-package com.recyclerviewmore.fragment;
-
-import com.recyclerviewmore.R;
-import com.recyclerviewmore.base.BaseFrgment;
-
-/**
- * zhangyao
- * 16/9/24
- * zhangyao@jiandanxinli.com
- */
-
-public class FragmentTwo extends BaseFrgment {
-    @Override
-    protected int getLayoutId() {
-        return R.layout.fragment_two;
-    }
-
-    @Override
-    public void init() {
-
-    }
-}
