--- app/src/main/java/com/recyclerviewmore/widget/LoadMoreListener.java
@@ -1,14 +1,0 @@
-package com.recyclerviewmore.widget;
-
-/**
- * zhangyao
- * 16/9/4
- * zhangyao@jiandanxinli.com
- */
-public interface LoadMoreListener {
-
-    /**
-     * loading more
-     */
-    void onLoadMore();
-}
