--- app/src/main/res/layout/app_bar_main.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.recyclerviewmore.MainActivity">
-
-    <android.support.design.widget.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="@style/AppTheme.AppBarOverlay">
-
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            app:layout_scrollFlags="scroll|enterAlways"
-            app:popupTheme="@style/AppTheme.PopupOverlay" />
-
-        <com.gigamole.library.navigationtabstrip.NavigationTabStrip
-            android:id="@+id/nts_bottom"
-            android:layout_width="match_parent"
-            android:layout_height="50dp"
-            android:layout_gravity="center"
-            android:background="@color/black_page_bg"
-            app:nts_active_color="#fff"
-            app:nts_color="@color/pink"
-            app:nts_corners_radius="1.5dp"
-            app:nts_factor="2.8"
-            app:nts_inactive_color="#81ffffff"
-            app:nts_titles="@array/titles"
-            app:nts_typeface="fonts/drugs.otf"
-            app:nts_weight="3dp" />
-
-    </android.support.design.widget.AppBarLayout>
-
-    <include layout="@layout/content_main"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@android:drawable/ic_dialog_email" />
-
-</android.support.design.widget.CoordinatorLayout>
