+++ app/src/main/res/layout/fragment_one.xml
@@ -4,11 +4,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.recyclerviewmore.widget.RecyclerViewUpRefresh
+    <redroid.widget.LoadMoreRecyclerView
         android:id="@+id/recyclerview"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-
-        ></com.recyclerviewmore.widget.RecyclerViewUpRefresh>
+        android:layout_height="match_parent" />
 
 </android.support.v4.widget.SwipeRefreshLayout>
