+++ loadmorerecyclerview/src/main/res/layout/view_footer.xml
